Änderung der Datenbank rückgängig gemacht, gab Konflikte wegen bereits gespeicherter Daten

+++ app/src/main/java/com/example/ep_prototyp/Design.kt
@@ -54,12 +54,10 @@ class Design : Fragment() {
     private fun längeBehaviourSpeichern(längeBehaviour:EditText) {
         mProfileDatabase.readRezeptData.observe(viewLifecycleOwner, Observer { data->
                 if(data.isEmpty()){
-                    mProfileDatabase.addRezept(Rezept(0,"weniger Essen",längeBehaviour.text.toString()))
+                    mProfileDatabase.addRezept(Rezept(0,"weniger Essen",längeBehaviour.text.toString().toInt()))
                 }else{
-                    mProfileDatabase.updateRezept(Rezept(1,data[0].rezeptBehaviour,längeBehaviour.text.toString(),data[0].rezeptCelebration,data[0].rezeptPrompt))
+                    mProfileDatabase.updateRezept(Rezept(1,data[0].rezeptBehaviour,längeBehaviour.text.toString().toInt(),data[0].rezeptCelebration,data[0].rezeptPrompt))
                 }
-
-
         })
 
     }
